+++ packages/dart/vault_data_manager/test/mocks/mock_iam_api_service.dart
@@ -1,3 +1,5 @@
+import 'package:affinidi_tdk_consumer_iam_client/affinidi_tdk_consumer_iam_client.dart'
+    as consumer_iam;
 import 'package:affinidi_tdk_iam_client/affinidi_tdk_iam_client.dart';
 import 'package:affinidi_tdk_vault/affinidi_tdk_vault.dart';
 import 'package:affinidi_tdk_vault_data_manager/src/services/vault_data_manager_shared_access_api_service_interface.dart';
@@ -77,5 +79,107 @@ class MockIamApiService extends Mock
         #onReceiveProgress: onReceiveProgress,
       }),
     ) as Future<Response<UpdateAccessOutput>>;
+  }
+
+  @override
+  Future<void> grantNodeAccessVfs({
+    required String granteeDid,
+    required List<String> nodeIds,
+    required Permissions permissions,
+    CancelToken? cancelToken,
+    Map<String, dynamic>? headers,
+    Map<String, dynamic>? extra,
+    ValidateStatus? validateStatus,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) {
+    return super.noSuchMethod(
+      Invocation.method(#grantNodeAccessVfs, [], {
+        #granteeDid: granteeDid,
+        #nodeIds: nodeIds,
+        #permissions: permissions,
+        #cancelToken: cancelToken,
+        #headers: headers,
+        #extra: extra,
+        #validateStatus: validateStatus,
+        #onSendProgress: onSendProgress,
+        #onReceiveProgress: onReceiveProgress,
+      }),
+    ) as Future<void>;
+  }
+
+  @override
+  Future<void> revokeNodeAccessVfs({
+    required String granteeDid,
+    required List<String> nodeIds,
+    CancelToken? cancelToken,
+    Map<String, dynamic>? headers,
+    Map<String, dynamic>? extra,
+    ValidateStatus? validateStatus,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) {
+    return super.noSuchMethod(
+      Invocation.method(#revokeNodeAccessVfs, [], {
+        #granteeDid: granteeDid,
+        #nodeIds: nodeIds,
+        #cancelToken: cancelToken,
+        #headers: headers,
+        #extra: extra,
+        #validateStatus: validateStatus,
+        #onSendProgress: onSendProgress,
+        #onReceiveProgress: onReceiveProgress,
+      }),
+    ) as Future<void>;
+  }
+
+  @override
+  Future<void> updateNodeAccessVfs({
+    required String granteeDid,
+    required List<String> nodeIds,
+    required Permissions permissions,
+    CancelToken? cancelToken,
+    Map<String, dynamic>? headers,
+    Map<String, dynamic>? extra,
+    ValidateStatus? validateStatus,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) {
+    return super.noSuchMethod(
+      Invocation.method(#updateNodeAccessVfs, [], {
+        #granteeDid: granteeDid,
+        #nodeIds: nodeIds,
+        #permissions: permissions,
+        #cancelToken: cancelToken,
+        #headers: headers,
+        #extra: extra,
+        #validateStatus: validateStatus,
+        #onSendProgress: onSendProgress,
+        #onReceiveProgress: onReceiveProgress,
+      }),
+    ) as Future<void>;
+  }
+
+  @override
+  Future<Response<consumer_iam.GetAccessOutput>> getNodeAccessVfs({
+    required String granteeDid,
+    CancelToken? cancelToken,
+    Map<String, dynamic>? headers,
+    Map<String, dynamic>? extra,
+    ValidateStatus? validateStatus,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) {
+    return super.noSuchMethod(
+      Invocation.method(#getNodeAccessVfs, [], {
+        #granteeDid: granteeDid,
+        #cancelToken: cancelToken,
+        #headers: headers,
+        #extra: extra,
+        #validateStatus: validateStatus,
+        #onSendProgress: onSendProgress,
+        #onReceiveProgress: onReceiveProgress,
+      }),
+    ) as Future<Response<consumer_iam.GetAccessOutput>>;
   }
 }